src/aggregation/core/requirements.txt
attrs==21.2.0
Automat==20.2.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
decorator==5.0.7
google-api-core==1.26.3
google-auth==1.30.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
googleapis-common-protos==1.53.0
grpcio==1.37.1
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
jsonpath-ng==1.5.2
lxml==4.6.3
packaging==20.9
parsel==1.6.0
ply==3.11
priority==1.3.0
Protego==0.1.16
proto-plus==1.18.1
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
queuelib==1.6.1
requests==2.25.1
rsa==4.7.2
Scrapy==2.5.0
service-identity==21.1.0
six==1.16.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
urllib3==1.26.4
w3lib==1.22.0
zope.interface==5.4.0
